+++ GetIntoTeachingApi/Controllers/CandidatesController.cs
@@ -92,50 +92,5 @@ namespace GetIntoTeachingApi.Controllers
 
             return NoContent();
         }
-
-        [HttpPost]
-        [Route("matchback")]
-        [SwaggerOperation(
-            Summary = "Perform a matchback operation, returning the match candidate id.",
-            Description = @"Attempts to matchback against a known candidate and returns the candidate id.",
-            OperationId = "MatchbackCandidate",
-            Tags = new[] { "Candidates" })]
-        [ProducesResponseType(StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
-        [ProducesResponseType(typeof(IDictionary<string, string>), StatusCodes.Status400BadRequest)]
-        public IActionResult Matchback([FromBody, SwaggerRequestBody("Candidate details to matchback.", Required = true)] ExistingCandidateRequest request)
-        {
-            request.Reference ??= User.Identity.Name;
-
-            if (!ModelState.IsValid)
-            {
-                return BadRequest(ModelState);
-            }
-
-            if (_appSettings.IsCrmIntegrationPaused)
-            {
-                _logger.LogInformation("CandidatesController - potential duplicate (CRM integration paused)");
-                return NotFound();
-            }
-
-            Candidate candidate;
-
-            try
-            {
-                candidate = _crm.MatchCandidate(request);
-            }
-            catch (Exception e)
-            {
-                _logger.LogInformation("CandidatesController - potential duplicate (CRM exception) - {Message}", e.Message);
-                throw;
-            }
-
-            if (candidate == null)
-            {
-                return NotFound();
-            }
-
-            return Ok(new CandidateMatchbackResponse(candidate));
-        }
     }
 }
