--- GetIntoTeachingApi/Models/CandidateMatchbackResponse.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using GetIntoTeachingApi.Models.Crm;
-
-namespace GetIntoTeachingApi.Models
-{
-	public class CandidateMatchbackResponse
-	{
-		public Guid CandidateId { get; set; }
-
-		public CandidateMatchbackResponse(Candidate candidate)
-		{
-            CandidateId = (Guid) candidate.Id;
-		}
-	}
-}
-
