+++ GetIntoTeachingApiTests/Controllers/CandidatesControllerTests.cs
@@ -134,74 +134,5 @@ namespace GetIntoTeachingApiTests.Controllers
             );
             _mockLogger.VerifyInformationWasCalled("CandidatesController - potential duplicate (CRM integration paused)");
         }
-
-        [Fact]
-        public void Matchback_InvalidRequest_RespondsWithValidationErrors()
-        {
-            var request = new ExistingCandidateRequest { Email = "invalid-email@", Reference = "Ref" };
-            _controller.ModelState.AddModelError("Email", "Email is invalid.");
-
-            var response = _controller.Matchback(request);
-
-            request.Reference.Should().Be("Ref");
-            var badRequest = response.Should().BeOfType<BadRequestObjectResult>().Subject;
-            var errors = badRequest.Value.Should().BeOfType<SerializableError>().Subject;
-            errors.Should().ContainKey("Email").WhoseValue.Should().BeOfType<string[]>().Which.Should().Contain("Email is invalid.");
-        }
-
-        [Fact]
-        public void Matchback_ValidRequest_ReturnsCandidateMatchbackResponse()
-        {
-            var request = new ExistingCandidateRequest { Email = "email@address.com", FirstName = "John", LastName = "Doe" };
-            var candidate = new Candidate { Id = Guid.NewGuid(), Email = request.Email, FirstName = request.FirstName, LastName = request.LastName };
-            _mockCrm.Setup(mock => mock.MatchCandidate(request)).Returns(candidate);
-            _mockAppSettings.Setup(m => m.IsCrmIntegrationPaused).Returns(false);
-
-            var response = _controller.Matchback(request);
-
-            request.Reference.Should().Be("Client");
-            var ok = response.Should().BeOfType<OkObjectResult>().Subject;
-            var matchbackResponse = (CandidateMatchbackResponse) ok.Value;
-            matchbackResponse.CandidateId.Should().Be((Guid)candidate.Id);
-        }
-
-        [Fact]
-        public void Matchback_MismatchedCandidate_ReturnsNotFound()
-        {
-            var request = new ExistingCandidateRequest { Email = "email@address.com", FirstName = "John", LastName = "Doe" };
-            _mockCrm.Setup(mock => mock.MatchCandidate(request)).Returns<Candidate>(null);
-            _mockAppSettings.Setup(m => m.IsCrmIntegrationPaused).Returns(false);
-
-            var response = _controller.Matchback(request);
-
-            response.Should().BeOfType<NotFoundResult>();
-        }
-
-        [Fact]
-        public void Matchback_CrmThrowsException_LogsAndReRaises()
-        {
-            var request = new ExistingCandidateRequest { Email = "email@address.com", FirstName = "John", LastName = "Doe" };
-            _mockCrm.Setup(m => m.MatchCandidate(request)).Throws(new Exception("Error"));
-            _mockAppSettings.Setup(m => m.IsCrmIntegrationPaused).Returns(false);
-
-            _controller.Invoking(c => c.Matchback(request))
-                .Should().Throw<Exception>()
-                .WithMessage("Error");
-
-            _mockLogger.VerifyInformationWasCalled("CandidatesController - potential duplicate (CRM exception) - Error");
-        }
-
-        [Fact]
-        public void Matchback_CrmIntegrationIsPaused_ReturnsNotFound()
-        {
-            var request = new ExistingCandidateRequest { Email = "email@address.com", FirstName = "John", LastName = "Doe" };
-            _mockAppSettings.Setup(m => m.IsCrmIntegrationPaused).Returns(true);
-
-            var response = _controller.Matchback(request);
-
-            response.Should().BeOfType<NotFoundResult>();
-
-            _mockLogger.VerifyInformationWasCalled("CandidatesController - potential duplicate (CRM integration paused)");
-        }
     }
 }
